# abled-app-backend/abledbackend-accounts_api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/users/{id}":{"get":{"tags":["Account"],"summary":"return the specific account with this id from our list.","operationId":"showAccountById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"put":{"tags":["Account"],"summary":"update the account with an specific id with the values introduced.","operationId":"updateAccountById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"delete":{"tags":["Account"],"summary":"delete the account with the specified id from our list.","operationId":"deleteAccountById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/users/mail":{"get":{"tags":["Account"],"summary":"return the specific account with an email introduced as a param, from our list.","operationId":"showAccountByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"put":{"tags":["Account"],"summary":"update the account with an specific email (introduced as a param) with the new values introduced.","operationId":"updateAccountByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"delete":{"tags":["Account"],"summary":"delete the account with an email introduced as a param, from our list.","operationId":"deleteAccountByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/users":{"get":{"tags":["Account"],"summary":"Return a list with all accounts","operationId":"showAllAccounts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}}},"post":{"tags":["Account"],"summary":"Create a new account","operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}}},"components":{"schemas":{"Account":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"lastName":{"type":"string"},"phoneCode":{"pattern":"^\\+[1-9]\\d{0,2}","type":"string"},"phoneNumber":{"pattern":"^\\d{1,4}-\\d{1,4}-\\d{1,10}$","type":"string"},"imageLink":{"type":"string"},"email":{"pattern":"^[\\w!#$%&’*+/=?`{|}~^-]+(?:\\.[\\w!#$%&’*+/=?`{|}~^-]+)*@(?:[a-zA-Z0-9-]+\\.)+[a-zA-Z]{2,6}$","type":"string"},"accountActive":{"type":"boolean"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"password":{"pattern":"^(?=\\w*\\d)(?=\\w*[A-Z])(?=\\w*[a-z])\\S{8,16}$","type":"string"},"rememberToken":{"type":"string"},"billingAddress":{"type":"string"}}}}}}